Rewrite code from javac code with for JDK-8272564
Some Dalvik and Art MVs does not support interface invokes to Object
members not explicitly defined on the symbolic reference of the
interface invoke. In these cases rewrite to a virtual invoke with
the symbolic reference java.lang.Object.

Conflicts:
Removed tests during cherry-pick. They used JDK-18 which is not
in third_party on the 3.2 branch.

Bug: 218298666
Change-Id: I38abdea8d9bbf4d004be1877373b58de161a55a7