Limit the size of pc2pc debug-info encoding.
This CL introduces pc2pc encoding points such that a given set of
methods is bounded by an expansion/non-sharing threshold.
The default value is 200.000 which is found to have very little
impact on size and avoids most regression time in dex2oat
speed-profile compilations.

For experimental purposes the defaults can be adjusted with:
 -Dcom.android.tools.r8.pc2pcOverheadThreshold=100000

Bug: b/231903117

Change-Id: I61296334a284d7c8461329eee5e8c4929b2b7f13