Rerun isolated split checker in final tree shaking
Running the checker in the first round of tree shaking guarantees that the application has no illegal accesses across feature split boundaries.

However, R8 may introduce errors as a result of optimizations, which would go unnoticed if we do not rerun the checker in the final round of tree shaking.

Bug: b/300247439
Change-Id: If23083b93b0fbc1774ef9806990c2a4ec1f77bc1