[KeepAnno] Introduce a reusable parser structure
The added parser differs from "declarations" by having a mapping of
property names to the logical parser cases. This allows sharing the
parsing of the annotation `TypePattern.name` with the external
properties such as `fieldType` and `methodReturnType`.

Follow-up CLs will extend the parser for composition and then replace
all the "declarations" and annotation visitors.

Bug: b/248408342
Change-Id: Ia95e0fa662f0b73c677c64c454c240d10644dcba