Make rewriting of new-array-filled better suited for class file output
This handles the situation where an array is populated
by an array of constants. When a new-array-filled of constants is
used by another new-array-filled the construction of the array
of constants is moved (copied) directly to the use site when possible.

This can avoid using a local slot for each of the new-array-filled
constants when generating to class file.

Bug: b/315753861
Change-Id: I0bae9c69d41ede7dc8aea622206f52b1fdabc9e9