Use the same set of unused resources as reported by the model
In the current implementation, the resource model will not report attr and styleable as unused, but in the removal phase these are still removed if we don't mark them as used.

The attributes are explicitly reachable from integer arrays in the R class, instantiated in clinit, which we explicitly ignore (in legacy mode)

This will also disable removal of actually unused attr resources in the optimized pipeline, I will do a follow up to fix that.

Bug: b/319802261
Change-Id: Ic242882bdc4307751ada82e09ca93094c4778133