Move LIR conversion into LirConverter
This also introduces a new LIR rewriting pass. This brings the LIR fully up to date by rewriting it with the current graph lens, but without converting the LIR to the output format. This allows building IR from the resulting LIR, which is a prerequisite for being able to build IR and apply the LensCodeRewriter in the final round of vertical class merging.

Bug: b/315445393
Change-Id: I2a0b13069d1a7c9b0b57af28947faa448e8ad028