Add workaround for test CtsPerfettoTestCases.HeapprofdJavaCtsTest#DebuggableAppOom
The cts test depend on the allocation in the following code staying in
debug mode:

public class JavaOomActivity extends Activity {
    @Override
    public void onCreate(Bundle state) {
        super.onCreate(state);
        new Thread(() -> {
            try {
                byte[] alloc = new byte[Integer.MAX_VALUE];
            } catch (OutOfMemoryError e) {
            }
        }).start();
    }
}

However, the local variable alloc has no local info, as it is not
observable in a debugger, so both D8 anbd R8 will remove it as dead
code in debug mode.

This adds a workaround keeping array allocations of size
Integer.MAX_VALUE to make the cts test work again.

Bug: b/322478366
Change-Id: I3c5c7c5aec062d5291c0c2ad90206fa839ffd208