Add regression test for non determinism in sorting classes for naming
When we sort the classes differently we will get a different name assignment.

The reason is that when we evaluate the javac generated bridge in A
and C, the order matters because the makeCommand method on A will
reserve the makeCommand name with arity 0.

The reservation is happening on the current frontier (which is
Foo). When we get to C this is reserved and we don't rename.

If we sort C first we did not yet reserve this.

Bug: b/326837090
Change-Id: I3e65311063aa21ff4a751838d9fbb64d2afa1794