Move API outlining and stubbing out of desugaring
This fixes the API outlining and stubbing to only be part of dexing
and also to be independent of desugaring.

Dispite being "outside" desugaring, The compilation still requires a
correct bootclasspath during dexing.

The temporary system property
`com.android.tools.r8.enableApiOutliningAndStubbing` is added to enable
this for CLI users.  The API already has a deprecated method to enable
this experimental feature until it is default enabled for the compiler.

Bug: b/324270842
Bug: b/333477035
Fixes: b/324500844
Change-Id: I51f4b7b623f5ac905e1de5382f792f0fcdf60fab