Unify first and final vertical class merging
This changes the first round of vertical class merging so that the vertical class merger now produces fully lens rewritten LIR code. This is consistent with the subsequent phases of vertical class merging.

For consistency with other optimizations, this CL also changes so that we no longer allow vertical class merging of classes that match -if rule preconditions.

As a result of this, there are no outstanding lenses when we enter the primary optimization pass.

Fixes: b/266049507
Fixes: b/318805362
Change-Id: I6a413d7658b8fed08a1c73e1dc8a8b19acc7972e