Fix handling of R class in optimized shrinking
With this change we will optimize R class fields, even when optimization is off.

This basically means that we will remove fields that is only written
in clinit, and have no other reads or writes.

To enable this, we now do deferred tracing in the first round of tree shaking.

Bug: b/336983087
Bug: b/325905703
Bug: b/287398085
Change-Id: I7dd72f1e9619340618d406876601cd1f2ea00f55