Update global signature configuration to unconditional disallow
The interpretation of pg keep rules must set the correct options on
keep-info and only the global disallow config should affect the judgment
for a given item.

This is in preparation for the native keep annotation evaluation
supporting disallow removal options even when the pg config does not
specify a -keepattributes for the kept attributes.

Bug: b/323816623
Change-Id: I236d6c88c65157df685e13a831bc6870d882df83