[KeepAnno] Introduce lattice for annotation keep info
This is in preparation for allowing keep annotations to selectively
keep subsets of annotations on items. The lattice can only assume
the top and bottom values, thus it is functionally equivalent to the
boolean value it replaces. Follow-up CLs will refactor the
determination what of annotations to keep and extend the lattice.

Bug: b/319474935
Change-Id: I4d7b8910ac4e222fa1942fd4dba733e73da0314b