Insert nop instructions to be step-out compatible with JVMs
For non-void invokes, the required pop instruction in the classfile
format would ensure that stepping out of such an invoke would return to
the same line as stepping in. In DEX no such instruction follows the
invoke and a nop is inserted to retain the same behavior.

This behavior is default disabled with this CL and setting the property `com.android.tools.r8.enableJvmCompatibleStepOutBehavior` is needed to enable it.

Bug: b/67671565
Change-Id: I85d4382d65f878057a3a2f34ab4dc903ac714dda