Disallow single caller inlining of already inlined methods
This fixes an issue in the second optimization pass when there is a call chain m1 -> m2 -> m3, where m2 is the single caller of m3.

If m2 is first inlined into m1 (e.g., due to meeting the simple inlining constraint), then m1 is now also a caller of m3. When processing m2 it is therefore invalid to single caller inline m3.

Note that this issue does not exist in the first optimization pass. Notably, if m2 is processed before m1, then m3 is correctly single caller inlined. If m1 is processed before m2, then m2 is not eligible for inlining into m1, since m2 is still unprocessed.

Fixes: b/343768871
Change-Id: I575b264dc25767560df4871a3d8387b3ce44f869