Always parse line and locals table
This removes the early check on parsing the line and local-variable
table. This unblocks the ability to determine the need for these
structures at a later point. In particular to include line tables and
method parameter names regardless of the `-keepattributes` settings.
In particular the latter is needed to support native keep annotations
which do not define properties in the pg configuration.

Fixes: b/260384637
Fixes: b/279555726
Bug: b/232212653

RELNOTES: R8 now encodes line-table mapping information for all
builds (effectively assuming -keepattributes LineNumberTable).
This was already the case for some builds targeting min-api 26
and above and is now the behavior for all builds. This can
cause minor size regressions for builds that used to discard
all line information and thus cannot be retraced. Due to the
compressed debug-info encoding used by R8 any such size
regression will be minor.

Change-Id: Ic0e8673b4e2f267d6eb162eb083267781f3c8ce3