[KeepAnno] Complete binding normalization of AST
This CL finishes the changes to the reader such that the AST is always
constructed in the normalized binding form. The AST still provides the
same API for "references" that are either binding-references or items so
as not to need changes to the extractor and native matcher. Follow-up
CLs will simplify theses consumers to assume the normalized AST.

Bug: b/343389186
Change-Id: I3e9a93ac09829f3f88ee0c4e87e49af01e064a56