Add support for always outlining APIs from extension libraries
This extends support for outlining extension APIs with the system property

  com.android.tools.r8.androidApiExtensionLibraries

for specifying one or more libraries (JAR files) with Android extension
APIs. Multiple extension libraries are passed by a comma separated list
of JARs.

All classes, methods and fields in the extension libraries are treated as
having an API level higher than any known API level thereby "simulating"
that these API might not be available no matter which min API level is
set for the compilation.

RELNOTE: Only mention this if AGP support is added as well.

Bug: b/326252366
Change-Id: I14bcf3a02a5ab41cbaff640357a16319ff5c9d37