Revert "[KeepAnno] Use native interpretation in R8"
This reverts back the API command to use rule extraction in the reader
instead of direct native interpretation. The keepanno tests do not use
the API call in this case such that they continue to test the "direct"
native interpretation.

This CL should be reverted to re-enable the native interpretation once
it is deemed safe for use in platform.

Bug: b/349276075
Change-Id: I02aec9c88417481685c123f27235b6d23a9eda6c