Support adding additional raw xml files to R8 resource shrinking
To ensure that we can parse all xml keep rules we introduce a new Kind for passing in rules. Rules are explicitly not output again, and only used for extracting keep rules.

This will ensure that we are compatible with the old resource shrinker
and additionally the setup can ensure that we will not hit cases where
aapt have removed duplicates (AGP can pass all files from the
individual res folders, even when there are name clashes)

Bug: b/287398085
Change-Id: I167c75e3b244d96870524e8a67141a22dd463d68