Declare new unused argument propagation states
An unused value state is planned to be used for a parameter P when an invoke is seen to the method of P, but where the argument to P won't be used inside the method.

After argument propagation, if a parameter P has an unused value, the parameter can be removed.

The unused states differ from the bottom states in that bottom for a parameter P encodes that the parameter does not take any value at runtime (meaning the method of P must be unreachable).

Change-Id: If7ff6966823786e917bf1d6de24a8be2dad59b79