Unsound, experimental removal of unused Composable arguments
For method parameters of Composable functions, this CL introduces a condition over the arguments to the Composable function, which evaluates to true if-and-only-if the argument can be skipped.

Since Composable method parameters generally always have uses, this CL ignores uses that are dominated by calls to skipToGroupEnd(), which is potentially unsound.

Bug: b/302281503
Change-Id: I3d7c2862f6aaabb9cd651540073cf8efa2391aa9