Add support for embedded rules in classpath and library providers
To expose other resources besides class files from classpath and library
providers the interface ClassFileResourceProvider is extended with the
method getDataResourceProvider() which is already present on
ProgramResourceProvider.

The method getDataResourceProvider() is an interface method with a
default implementation, so it is backwards compatible with existing
ClassFileResourceProvider implementations except if they have a
method with this signature already.

System property com.android.tools.r8.readEmbeddedRulesFromClasspathAndLibrary
is currently requres to enable this.

Bug: b/289087274
Change-Id: I7e17a05d155560a5ca0a7d93e2d0615f88984d47