Inlining of constructors with final field assignments
The optimization uses VarHandle.storeStoreFence() to preserve semantics after inlining. This was added to android.jar in API 34 but from testing seems to be present since API 28 (yet to be confirmed).

This CL currently does not unset the final flags of fields that are no longer only assigned in constructors. Since we are not at a join point at the time of inlining we will need to postpone this until the next join point to ensure determinism.

The optimization is still disabled.

Bug: b/278964529
Change-Id: I7733a217aff7e5d560667e7418137fb0434929de