Split live ranges for invoke/range
This changes the register allocation strategy for dealing with invoke/range. Unlike previously, this no longer creates alias SSA values using move instructions for all arguments to an invoke/range instruction. Instead, this splits the live ranges of values that are used in invoke/range instructions upfront. Similar to the previous strategy, when processing the next unhandled live intervals, we first assign registers to all subsequent split live intervals for the same value (and their linked values) that span an invoke/range instruction.

One key advantage is that this avoids the need for alias SSA values, which has the negative side effects that the same SSA value occupies two registers and that the register allocator did not leverage that a given value was known to be in multiple registers. As a concrete example, the register allocator would occasionally move a value to a low register prior to an invoke/range instruction, and then after the invoke/range add a redundant move for the same value to a low register, since the existing value in a low register would belong to a different SSA value.

This change identifies and solves the following issues:

* When the same value is used as more than one argument to the same invoke/range, alias SSA values are introduced since each live intervals can (currently) only have one register (two consecutive if wide).

* Unlike previously, the live intervals that are used to assign registers to invoke/range instructions can be live out of the invoke/range instruction. As a result, this may allocate all 4 bit or 8 bit registers in a way that there are no available registers after the invoke/range (effectively running out of registers). To mitigate this, we always block the lower most register(s) when the invoke/range has an out-value.

* When allocating registers for methods with <=16 registers this no longer pins arguments in their input register (except for the receiver), leading to lower register pressure.

* When allocating registers for methods with <=16 registers that have an invoke/range instruction and use catch handlers, register allocation now needs to block a dedicated move-exception register throughout the method, similarly to in 8 bit and 16 bit register allocation.

* A bug in the removeUnneededMovesOnExitingPaths optimization (debug mode only).

This saves 6-7% of all moves in Compose sample benchmarks and +10% of all moves in Composable functions.

Bug: b/302281605
Change-Id: I172640414ac94f79d170da7005daf80827f4562b