Allow using move exception register for values not live to catch
This records on each live intervals whether the given live interval is live at entry to a block that starts with a move-exception instruction.

All values that are not live at entry to a move-exception block do not need to check for overlap with move exception live intervals.

This improves build speed but also size, since this means that the move-exception split live intervals will be able to reuse the dedicated move-exception register. In other words, this fixes the issue that the compiler would almost always insert a (often) redundant move immediately after each move-exception instruction.

Change-Id: I22e0c6b040dd49c115680034adc148bc35ca72bd