Don't use reflective lookup for calling command builder code
This will compile a set of java files against TOT, allowing us to
directly use the api instead of reflective lookup.

We put this on classpath for the compiledump invocation, and catch any
missing class or method errors.

New files in this directory are automatically compiled, but one should
ensure that functionality that are added in different versions of R8
are added to different files (to avoid issues loading the class for
old functionality)

This only enables this for resource shrinker, I will change the rest
if we agree on this approach.


Bug: b/377606434
Change-Id: I6db890440de6c188f5da15c55723fc1579523c17