Introduce disabled optimization ListIterationRewriter
This is to rewrite iterator-based for-each loops to indexed loops.

Adds a test, InternalOptions flag, and initial optimization.

It does not kick in when List subclass is not known.

For Chrome, this converts 1010 out of 1731 possible loops. The
remaining 721 are those where the type is not ArrayList or
ImmutableList (e.g. most are of type "List").
The DEX size shrank by 300 bytes.

Follow-ups:
 * Enable the optimization
 * Add more allowlisted list types (e.g. kotlin lists)

Bug: b/145280859
Change-Id: Ifb63ad5295207f6309060db9cd6d4be58a46d4d7