Enable 8 bit register allocation refinement pass
This reduces the number of moves in JetNews by 0.65%.

When compiling JetNews with R8 we perform register allocation for 17107 methods. We hit 8 bit register allocation in 624 methods (3.5%). For 115 of these methods (0.7%), we run the 8 bit refinement pass. For 20 of these methods (0.1%), the 8 bit refinement pass fails, meaning we need to rerun the original 8 bit register allocation pass.

Bug: b/374266460
Change-Id: Ib0acb0d19f19480f6c99fc8f111a516bfc24767f