Require instance fields are kept when instantiating classes using Unsafe
This effectively reverts 6250eeeafa351b13ef371e8a67f5b8c641459e1b.

The reason for the revert is that this causes +1% regressions in dex size. When sun.misc.Unsafe is used to instantiate a class without an instance initializer, the previous behavior can be restored by keeping the instance fields of the instantiated class. See also DefaultFieldValueAnalysisWithKeptSubclassTest.

Bug: b/379034741
Change-Id: Ie3795040413e888691ae1316e1878da9f77f6707