Pass resource provider/consumer/config to R8 partial configuration
Add test that this works in the simple case where the R class is in R8 compiled code.
In this simple case, the R class field reachable from D8 comiled code
is explicitly kept by the trace reference rules, and hence we keep the
resource.

Add test where this does not work for cases where the R class resides in the D8 part.

Bug: b/388746233
Change-Id: I518d2cffae2cf0a3ed887e1b99b0c8a0d5120cef