Avoid parsing app more than once in R8 partial
This removes the dumping of the input in R8 partial to a zip and ensures that the read classes in the R8 partial partitioning step are retained in memory and passed directly to the D8 and R8 compilations, so that we do not parse the input application more than once.

This still retains the dexing and desugaring steps as two separate compilations. Follow up work will merge these into a single D8 compilation.

Change-Id: I5b4bf894eb5814f4e9a429868b952b5ee10a19c2