Wrap desugaring collection in D8
This refactors the desugaring collection in D8 so that a wrapper around the desugaring collection returns the underlying desugaring collection given the class/method that is about to be desugared.

This will allow returning different desugaring collections based on the class/method being desugared, which will be used in the D8 step of R8 partial, to apply different desugarings to the D8 part and R8 part.

Bug: b/391572031
Change-Id: I050cf51a40df816313b94f17ccd5eb39a8c688a7