Support multiple keep annotation namespaces
With this change the keep annotations can be in either the current
com.android.tools.r8.keepanno namespace or in the new
androidx.annoataion.keep namespace. The compiler will treat both
namespaces the same, and annotations from both namespaces can be
present at the same time.

The support for the new namespace is needed to be able to release the
keep annotation library and the support for both is required for
migrating the use of keep annotations in android platform.

Bug: b/392865072

Change-Id: I173c630269731037e865dbd4b55800862647f563