Remove references to pruned methods in argument propagation
This fixes a bug where argument propagation proves that a direct method is dead. Argument propagation then removes the method from the app, but does not forcefully remove all calls to the method (which indirectly have been proved to also be dead).

This CL makes an update to argument propagation so that the removal of the dead methods is deferred until all calls to these methods have been removed.

Change-Id: Id71255467e2daae1395550e5d5bb15870d4fd396