Add support for LIR rewriting using custom lens
In order to implement library desugaring as a LIR-to-LIR transformation using a graph lens, it is desirable not to commit a library desugaring graph lens to the graph lens chain, since the graph lens only represents a rewriting of the code (i.e., it is not required for building the mapping file).

The LirLensCodeRewriter currently applies the lenses starting from the current (unapplied) graph lens to the code lens. This CL adds support for applying an uncommitted graph lens.

Change-Id: I514207174be6e82aed8dacba4717a7a0edf975e9