Ensure that resources are kept if there are keep rules on R class fields
We would previously not keep the entries if the R class values was
embedded as static field values (no clinit)

The test is rewriting the generated R class to ensure that the field
has a static value instead of being initialized in clinit.

Note that this can only trigger if there are no references to the R class at all.

Change-Id: Id7d52c7e8af3b2463e6dc48d420b8da9fd3f2a83