Add support for simple lir based resource value extraction
In R8Partial we will need to find all resource references in the D8
code, both resources referenced in the D8 part itself and resources
referenced in the R8 part. This is currently doing a full traversal of
all code, but if we make the choice of always using optimized resource
shrinking in R8 partial we can change this to only traverse the R
classes in the D8 part (since optimized resource shrinking requires
non final R class fields all resources referenced in the R8 part will
be automatically traced in the enqueuer.

Bug: b/400935182
Change-Id: Ia76a8899100869e1e6375913b48aa125b8ba71df