Fix incorrect comparison of virtual registers in rematerialization
This fixes a register allocation bug that only manifests when methods require more than 256 registers.

The const instructions in DEX have an 8 bit register constraint. When determining if a value can be rematerialized, the register allocator therefore finds the max non-spilled register for the live intervals of the value to check if it fits in 8 bits.

However, in the computation of the max non-spilled register, the register allocator failed to consider that the virtual registers 0...N where N=number of arguments are larger than registers N+1...M where M is the max virtual register.

Bug: b/408438353
Change-Id: I6c9d8c0d99bdb77a1678411995df5579f91b9ef3