Enable container DEX from min API 36 (BAKLAVA)
This also enables reading of DEX container format without having to
specify system property com.android.tools.r8.dexContainerExperiment

The system property com.android.tools.r8.dexContainerExperiment can
still be used to enable container DEX, but a warning will be emitted
if the min API level is below 36.

With this the recommended way of enabling container DEX is to specify
a min API level of 36 or higher.

Fixes: b/407778841
Bug: b/409353491
Bug: b/249922554
Change-Id: Id95479476c146f53280df3bffa5d676a8689bf1c