Allow blocks with catch handlers to have more than one normal successor
We currently have an invariant that all blocks with catch handlers end in a goto/return/throw instruction. The compiler does not seem to rely on this anywhere.

This CL relaxes the SSA validity checks so that blocks with catch handlers can end in any jump instruction (e.g., if and switch). This also extends the redundant block removal to account for this relaxation, since this change makes it possible to eliminate more gotos in the IR.

Change-Id: I0eee9d8d18af741924542ce4924a731612496e05