Simplify field access info APIs
This changes the APIs offered by the FieldAccessInfo so that they become less general purpose, which allows more efficient implementation.

As an example, all uses of `isWrittenOnlyInMethodSatisfying` is replaced by a new `isEffectivelyFinal` API.

As a first milestone, this CL removes all uses of the read contexts, which makes it possible to clear all read contexts after tree shaking. As a result, lens rewriting of the FieldAccessInfoCollection become more efficient.

Change-Id: I4f443290bbd5960c7f22a6a6b81cc0aa4dcf5efd