Disable propagation of non-package private final method states
Final methods cannot be overridden and therefore the top-down propagation of argument information does not need to push down states for final methods.

This still forwards states for package-private final methods. The reason for this is that there may exist a method with the same signature as a package-private final method in a subclass in a different package, since this is not an override of the package-private method.

This CL also extends timing tracking to the top-down propagation.

Bug: b/424006396
Bug: b/422947619
Change-Id: Ib8a6d4c8705aa950bdba57c6bae1a1b94d53c0d6