Parallelize NoClassInitializerCycles policy
This changes the policy so that prior to running the policy (in the preprocessing), the tracing is run concurrently starting from each merge group. If a group is found to be safe for merging, it is "committed". Class initializer cycles should be rare even after merging, and thus this should lead to a high level of parallelization.

Bug: b/422947619
Change-Id: I09a2a54e70087e7517ac5a4ca5536849d0644b31