Change the default emitted source file to 'r8-map-id-<MAP_ID>'
RELNOTES:

The default emitted source file attribute for a class changes from
'SourceFile' to 'r8-map-id-<MAP_ID>' when retracing is required (i.e.,
when either obfuscation or optimization is enabled).

Given an obfuscated stack trace this makes it possible to extract the id
of the mapping file that is required for retracing, which can be used to
implement automated retracing of stack traces in logcat.

If a custom source file attribute is used (-renamesourcefileattribute)
then this continues to take precedence.

In ProGuard compatibility mode, this only takes effect if the SourceFile
attribute is *not* kept (-keepattributes SourceFile). Apps that use
ProGuard compatibility mode and want to include the mapping file id in
their stack traces should remove -keepattributes SourceFile (or migrate
to R8 full mode).

This also changes the map id from being a 7 character prefix of the map
hash to being the full map hash.

Bug: b/424303069
Change-Id: I933b0dee01568bdfe4099264ade168ad2dca33ba