Optimize subtyping checks in argument propagator
This optimizes >= checks between two type elements in the argument
propagator into >= checks between a type element and a class type.

In many cases this avoids the need to: (1) Convert the class type to a
class type element (which involves synchronization), and (2) Compute the
least upper bound of the two class type elements.

Bug: b/422947619
Bug: b/424006396
Change-Id: I11542a5308667869961cfa6c5081f367e661c966