Optimize supertype traversal during subtype checks
This splits AppInfoWithClassHierarchy#traverseSuperTypes into two methods traverseSuperTypes and traverseSuperInterfaces, where the latter only gives callbacks for the super interfaces.

Moreover, the traverseSuper* methods are extended to support stopping the upwards class hierarchy traversal from a given class (specified by a predicate).

This is used to optimize subtype checks in the following way:

* If the superclass of interest is an interface, then the new traverseSuperInterfaces is used to skip the initial superclass chain traversal, which is known in advance not to find the superinterface of interest.

* If the superclass of interest is a program class or interface, then the upwards class hierarchy traversal is stopped at the class/library boundary, so that classes and interfaces in the library will not be considered.

Bug: b/422947619
Bug: b/424006396
Change-Id: Id4bf5966b56fd10a2fcb55167bb6f1270ec46bce