Refactor position remapper to better facilitate concurrency
When the line number optimization runs, a single stateful  ClassPositionRemapper is used, which tracks the following state:

  DexMethod previousMethod = null;
  int previousSourceLine = -1;
  int nextOptimizedLineNumber = 1;

This CL splits the position remapper into an app position remapper, a class position remapper, and a method position remapper. A stateful class position remapper can be created from the app position remapper, and a stateful method position remapper can be created from a class position remapper.

This makes it possible for different threads to create their own state for the classes and methods they process. This is outside the scope of the current CL.

Bug: b/422947619
Change-Id: I3273a235c6a237a8f1fb7d9667815ce1f9665bd1