Optimize deferred annotation processing in tree shaking
Instead of processing all deferred annotations when reaching an Enqueuer fixpoint, this changes the deferred annotation processing to run on-the-fly when annotation classes become live and when classes/fields/methods have their keep info changes.

This also changes the annotation tracing so that it enqueues worklist items. This makes it trivial to parallelize annotation tracing in the future since the worklist is thread safe, if needed.

Bug: b/422947619
Change-Id: I99d8af2051d45e8a3f9458ce370a18a606b49ef6