Ignore -dontusemixedcaseclassnames and always apply it
Never generate mixed case class and package names which collide if
compared case-insensitive. This is the safest, as this can cause issues
when extracting R8 generated code on file systems which are not
case sensitive (mainly Windows).

This also applies to names from -classobfuscationdictionary and
-packageobfuscationdictionary (which was already the case).

The output can still contain classes and packages with colliding
mixed case names if these where present in the input.

Member names (methods and fields) can still be in mixed case (this has
not changed, as -dontusemixedcaseclassnames did not affect member
names).

RELNOTES: R8 will no longer generate package and class names which
collide if compared case-insensitive. Previously this had to be enabled
with the option -dotnusemixedcaseclassnames which no longer has any effect
and will be silently ignored if passed to R8 as the effect always applies.

Fixes: b/434863985
Change-Id: I9f21fc3b04908d4eb197dfa4b648919451c167e2