Support @UsesReflectionToConstruct for R8 native interpretation
This extend the keep edge building with a constraint to keep the
kotlin.Metadata annotation.

For keep rule generation this was not required, as extraction to keep
rules does not support selecting specific annotation to keep for each
kept class.

Actually keeping kotlin.Metadata ialos relies on the explicitly added
keep rule `-keep kotlin.Metadata { *; }` currently added in both the
test setup and by the embedded keep rules in the Kotlin reflect library.

Bug: b/392865072
Change-Id: I39a0aee3964616fac7564113ffd477835e8fc64b