Outline construction for no-argument throw blocks
This introduces a pass that during D8 IR conversion finds blocks that
end in a throw, and tries to build an outline from the throw instruction
by recursively building up the outline method.

The pass:
- handles blocks on the form `throw new T()`
- builds IR that is converted to LIR on successful outline creation
- efficiently computes outline equivalence on LIR by checking equivalence of the underlying instruction byte arrays and the LIR constant pool

It is left for future work to materialize the outlines and rewrite the
code into calling the outlines.

Bug: b/434769547
Change-Id: I436efa1130fd0c45a9c3ede6ce07920649ba7293