Add outline markers at the entry point of outline extraction
This introduces a new synthetic IR instruction named ThrowBlockOutlineMarker, which references a concrete outline.

The pass that extracts outline candidates now inserts a ThrowBlockOutlineMarker instruction at the entry of each extracted outline. Later, when it has been established which outline candidates to materialize and which ones to discard, the outline optimization can easily find the program points where to insert a call to the materialized outline methods.

Since the ThrowBlockOutlineMarker instructions need to be processed later, all methods where such an instruction is inserted will be converted to LIR rather than DEX. This allows efficient reprocessing of these methods. This CL therefore also adds support for the ThrowBlockOutlineMarker instruction in the LIR code representation.

Bug: b/434769547
Change-Id: I39ad41c8c5e2ffd61365023379d6d2af72b77040