Extend throw block outliner to exceptions with arguments
This adds support for passing arbitrary values to the outlined exception. The outliner captures these values as arguments of the outlined method.

This also adds support for outlining uses of StringBuilder.

To keep track of which values from a given outline user must be passed as arguments to the outline method, this CL extends the ThrowBlockOutlineMarker to have as in-values the list of SSA values that are arguments to the outline method.

Bug: b/434769547
Change-Id: I0ca6343daf412347be4194ec8965e01df372651e