Add retracing benchmarks for varying sizes of mapping files
This adds a new script in tools for synthesizing mapping files of a given size from an existing mapping file.

The script is used to generate the following new mapping files in third_party/retrace_benchmark for benchmarking.

* r8lib-250MB.jar.map
* r8lib-500MB.jar.map
* r8lib-750MB.jar.map
* r8lib-1GB.jar.map

This CL also adds the following mapping files in the same directory, which has been built from the above mapping files using com.android.tools.r8.retrace.Partition.

* r8lib-250MB.jar.map.partition
* r8lib-500MB.jar.map.partition
* r8lib-750MB.jar.map.partition
* r8lib-1GB.jar.map.partition

Finally, this adds 12 new benchmarks to go/retraceperf. For each mapping file of a given size, a retrace benchmark is added that:

* Reports the avg retrace time for retracing when the mapping file is read from disk
* Reports the avg retrace time for retracing when the mapping file is kept in memory
* Reports the avg retrace time for retracing when the partitioned mapping file is used (read from disk)

Fixes: b/438901135
Change-Id: Ic2cb10a9dd8f77c7c003be69514ac709a0ad7334