Parallelize line number optimizer
This CL makes the observation that LineNumberOptimizer#run iterates over the app classes in non-deterministic order and commits the result of each class to the classNamingBuilder, meaning it can easily be parallelized.

In order to parallelize this the parsing of Kotlin source debug extensions is performed ahead of time.

Another important change is that this CL changes the CfLineToMethodMapper from re-parsing all classes from the program resource provider to only re-parsing the Kotlin classes with inline functions.

Bug: b/422947619
Bug: b/220999985
Change-Id: Ibe63f75fb2e13f0b3457f352b07fc43de5248251