Archive and aggregate perf try runs
When running jobs on the 'perf' bot, the option --patch-set should be set for try jobs. When this option is set, we now archive the results to gs://r8-perf-results/try/.

When aggregating the data in the bucket as part of creating a single json file for go/r8perf, we now look for hashes in gs://r8-perf-results/try/ and treat include these as try runs.

The aggregated try runs are archived to their own json file at gs://r8-perf-results/r8_benchmark_try_data.json.

Change-Id: Id7acc6ed7a021b22dbeb3616d3a9b84672c88713