Support for running throw block outliner in debug mode
It is a prerequisite for running the outliner in debug mode that the system property com.android.tools.r8.throwblockoutliner.forcedebug is set.

This is not the default since outlining in debug mode may strip debug locals. It also makes it impossible to step in a throw block that has been outlined.

Bug: b/434769547
Change-Id: I92b4b8f67d43582bd67abc1c34e48ed81009063b