Remove support for generating L8 rules from D8/R8
RELNOTES:

This removes support for generating keep rules for L8 from D8 and R8. Clients should instead use TraceReferences for this purpose.

More specifically, the methods D8Command.builder.setDesugaredLibraryKeepRuleConsumer and R8Command.Builder.setDesugaredLibraryKeepRuleConsumer are removed, and the support for --desugared-lib-pg-conf-output is removed from the command line options of D8 and R8.

Change-Id: Ia7f03c36d4ed63d93c8ee0f29092212ee6d2036e